refactor: 픽스쳐 위치 이동
