+++ backend/src/test/java/com/digginroom/digginroom/controller/MemberJoinControllerTest.java
@@ -2,6 +2,7 @@ package com.digginroom.digginroom.controller;
 
 import static org.hamcrest.Matchers.equalTo;
 
+import com.digginroom.digginroom.TestFixture;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import org.junit.jupiter.api.Test;
