+++ backend/src/test/java/com/digginroom/digginroom/service/MemberServiceTest.java
@@ -1,6 +1,6 @@
 package com.digginroom.digginroom.service;
 
-import static com.digginroom.digginroom.controller.TestFixture.파워;
+import static com.digginroom.digginroom.TestFixture.파워;
 import static com.digginroom.digginroom.domain.Genre.DANCE;
 import static com.digginroom.digginroom.domain.Genre.ROCK;
 import static org.assertj.core.api.Assertions.assertThat;
